backend/requirements.txt
django==5.2.6
djangorestframework==3.15.2
pytest-django==4.9.0
uvicorn
psycopg2-binary==2.9.10
dataclasses-json==0.6.7
djangorestframework-dataclasses==1.3.1
djangorestframework-simplejwt==5.3.1
psutil==7.1.3
pillow==12.0.0